+++ kernel/include/kernel/mem/mem.h
@@ -1,9 +1,11 @@
 #ifndef MEM_H
 #define MEM_H
     #include <kernel/multiboot_32.h>
+    #include <stddef.h>
+    
     void allocate_buffer();
     void parse_available_mem();
     uint32_t* get_page();
     void free_page(uint32_t *page);
-    int get_num_free_pages();
+    size_t get_num_free_pages();
 #endif
